requirements.txt
annotated-types==0.6.0
anyio==4.1.0
black==23.11.0
certifi==2023.11.17
click==8.1.7
dependency-injector==4.41.0
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
platformdirs==4.1.0
pydantic==2.5.2
pydantic_core==2.14.5
python-dotenv==1.0.0
python-telegram-bot==20.7
six==1.16.0
sniffio==1.3.0
tomli==2.0.1
typing_extensions==4.8.0
